% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{add_date_from_period_id}
\alias{add_date_from_period_id}
\title{Parses the period_id as a date for plotting}
\usage{
add_date_from_period_id(df, interval)
}
\arguments{
\item{df}{a dataframe with at least a period_id column}

\item{interval}{"M" or "Q" right now, in practice  `interval <- get_interval(vintage, con)`}
}
\value{
teh same dataframe with an added period column
}
\description{
Taking the datapoints table or any dataframe with a period_id column,
the function parses it as a date e.g. 2022M02 becomes 2022-02-01.
Currently works for months and quarters. Gets the interval from the database.
}
