% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{create_selection_excel}
\alias{create_selection_excel}
\title{Prepare excel file for series selection}
\usage{
create_selection_excel(df, outfile = "db_series", overwrite = TRUE)
}
\arguments{
\item{df}{dataframe, probably  from \link[UMARaccessR]{get_all_series_wtable_names}}

\item{outfile}{name of excel file to save to}

\item{overwrite}{- whether or not to overwrite previous file}
}
\value{
nothing - side effect is saving to an excel file.
}
\description{
Takes the dataframe with all the series produced by \link[UMARaccessR]{get_all_series_wtable_names}
and creates an excel file with these series and a prepared sheet for creating a selection of
series.
}
