% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_all_series_wtable_names}
\alias{get_all_series_wtable_names}
\title{Get all series in the database}
\usage{
get_all_series_wtable_names(con)
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
dataframe with a row for each series
}
\description{
Extracts a dataframe with all the series in the database joined together
with their table names and units. Used in \link[UMARaccessR]{create_selection_excel}. Cannae be
bothered with testing.
}
