% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_data_points}
\alias{get_data_points}
\title{Get and prep the datapoint table for a single vintage}
\usage{
get_data_points(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
a three column dataframe with the period_id, period as date and value for all
the datapoints in that vintage.
}
\description{
Gets the data periods and values for a single vintage, converting the periods
to real dates for easier plotting.
}
