% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_data_points_from_vintage}
\alias{get_data_points_from_vintage}
\title{Get and prep the datapoint table for a single vintage}
\usage{
get_data_points_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
a two column dataframe with the period_id, and value for all
the datapoints in that vintage.
}
\description{
Gets the data periods and values for a single vintage.
}
\details{
Deprecated: Use `\link[UMARaccessR]{sql_get_data_points_from_vintage}` instead.
}
