% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_initials_from_author_name}
\alias{get_initials_from_author_name}
\title{Get initials from author name}
\usage{
get_initials_from_author_name(name, con)
}
\arguments{
\item{name}{author's name}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character initials
}
\description{
Get initials from author name
}
