% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_interval_from_series}
\alias{get_interval_from_series}
\title{Get table interval from series ID}
\usage{
get_interval_from_series(series, con)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{vintage}{numeric id of vintage}
}
\value{
character vector of interval id
}
\description{
Get the interval id from the series table, which is alphanumeric e.g. M for monthly etc.
}
