% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_interval_from_vintage}
\alias{get_interval_from_vintage}
\title{Get table interval from vintage ID}
\usage{
get_interval_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of interval id
}
\description{
Joins the vintage table with the series table to get the
interval id, which is alphanumeric e.g. M for monthly etc.
}
\details{
Deprecated: Use `\link[UMARaccessR]{sql_get_interval_from_vintage}` instead.
}
