% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_last_period_from_vintage}
\alias{get_last_period_from_vintage}
\title{Get the most recent period for whic a vintage has data}
\usage{
get_last_period_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character of period id e.g "2022M03"
}
\description{
The vintage must have an actual non null value for that period.
}
