% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_max_category_id_for_source}
\alias{get_max_category_id_for_source}
\title{Get the largest existing category id for a given source}
\usage{
get_max_category_id_for_source(source_id, con)
}
\arguments{
\item{source_id}{numeric source id}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric id
}
\description{
Get the largest existing category id for a given source
}
