% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_metadata_from_author_name}
\alias{get_metadata_from_author_name}
\title{Get structural metadata for individual UMAR author}
\usage{
get_metadata_from_author_name(author_name, con)
}
\arguments{
\item{author_name}{author's name}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character initials
}
\description{
This function is a leftover from a dead end approach, but might still come
in handy. it takes an author name and grabs most of the structural metadata
for that authorćs series in the database.
}
