% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_series_id_from_series_code}
\alias{get_series_id_from_series_code}
\title{Get series id from series code}
\usage{
get_series_id_from_series_code(series_code, con)
}
\arguments{
\item{series_code}{series code value}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric id
}
\description{
Get series id from series code
}
