% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_series_ids_from_table_id}
\alias{get_series_ids_from_table_id}
\title{Get series ids from table id}
\usage{
get_series_ids_from_table_id(id, con)
}
\arguments{
\item{id}{numeric table id}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric ids
}
\description{
Deprecated: Use `\link[UMARaccessR]{sql_get_series_ids_from_table_id}` instead.
}
