% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_series_name_from_series_code}
\alias{get_series_name_from_series_code}
\title{Get series name from series code}
\usage{
get_series_name_from_series_code(series, con)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of series name.
}
\description{
Get the series name_long.
}
