% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_series_name_from_vintage}
\alias{get_series_name_from_vintage}
\title{Get series name from vintage ID}
\usage{
get_series_name_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of series name.
}
\description{
Joins vintage table with series to get the series name_long.
}
\details{
Deprecated: Use `sql_get_series_name_from_vintage()` instead.
}
