% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_tab_dim_id_from_table_id_and_dimension}
\alias{get_tab_dim_id_from_table_id_and_dimension}
\title{Get dimension id from table id and dimension name}
\usage{
get_tab_dim_id_from_table_id_and_dimension(id, dimension, con)
}
\arguments{
\item{id}{numeric table id}

\item{dimension}{character string with dimension name}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric code
}
\description{
Deprecated: Use `\link[UMARaccessR]{sql_get_tab_dim_id_from_table_id_and_dimension}` instead.
}
