% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_table_name_from_series}
\alias{get_table_name_from_series}
\title{Get table name from series ID}
\usage{
get_table_name_from_series(series, con)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of table name.
}
\description{
Lookup the table name from the series ID
}
\details{
Deprecated: Use `\link[UMARaccessR]{sql_get_table_name_from_series}` instead.
}
