% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_table_name_from_vintage}
\alias{get_table_name_from_vintage}
\title{Get table name from vintage ID}
\usage{
get_table_name_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of table name.
}
\description{
Joins the vintage table with the series table to get the
table id and then looks up the name in the table table.
}
