% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_unit_from_series}
\alias{get_unit_from_series}
\title{Get unit name from series id}
\usage{
get_unit_from_series(series, con)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of unit name.
}
\description{
Joins vintage and series tables to get the id of the unit,
which is looked up in the unit table to get its name.
}
