% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_unit_from_vintage}
\alias{get_unit_from_vintage}
\title{Get unit name from vintage id}
\usage{
get_unit_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character vector of unit name.
}
\description{
Joins vintage and series tables to get the id of the unit,
which is looked up in the unit table to get its name.
}
