% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_unit_id_from_unit_name}
\alias{get_unit_id_from_unit_name}
\title{Get unit id from name}
\usage{
get_unit_id_from_unit_name(name, con)
}
\arguments{
\item{name}{character name of unit}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric id
}
\description{
Get unit id from name
}
