% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_vintage_from_series}
\alias{get_vintage_from_series}
\title{Get vintage id from series id}
\usage{
get_vintage_from_series(series, con, date_valid = NULL)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{date_valid}{date when the vintage was valid if none is given most recent
i.e. currently valid vintage is returned.}
}
\value{
numeric vintage id.
}
\description{
Gets the vintage based on the series ID, by default the most recent one,
otherwise the one that was valid on the date passed with `date_valid`.
}
