% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{prep_single_line}
\alias{prep_single_line}
\title{Prepare data needed for a univariate line chart}
\usage{
prep_single_line(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
a list with the dataframe with values, period_ids and periods as the
first element, a character unit name as second, and the row wrapped main and
subtitles (default 100 chars max 3 lines).
}
\description{
Given the vintage id and a connection to the database, this function
gets the datapoints, and the unit, prepares the titles.
}
