% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_all_series_wtable_names}
\alias{sql_get_all_series_wtable_names}
\title{Get all series in the database}
\usage{
sql_get_all_series_wtable_names(con, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with a row for each series containing table and unit information
}
\description{
Get all series in the database
}
\details{
Extracts a dataframe with all the series in the database joined together
with their table names and units.
}
