% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_category_id_by_name}
\alias{sql_get_category_id_by_name}
\title{Get latest datapoints for table with vintages and all}
\usage{
sql_get_category_id_by_name(con, cat_name, source_id, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{cat_name}{Name of category}

\item{source_id}{numeric source id}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with table IDs and codes
}
\description{
Get latest datapoints for table with vintages and all
}
