% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_date_published_from_vintage}
\alias{sql_get_date_published_from_vintage}
\title{Get publication date from vintage ID}
\usage{
sql_get_date_published_from_vintage(con, vintage_id, schema = "test_platform")
}
\arguments{
\item{con}{Database connection object}

\item{vintage_id}{Integer vintage identifier}

\item{schema}{Character string specifying the database schema}
}
\value{
POSIXct timestamp with CET timezone
}
\description{
Get publication date from vintage ID
}
