% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_dimension_id_from_table_id_and_dimension}
\alias{sql_get_dimension_id_from_table_id_and_dimension}
\title{Get dimension ID for table ID and dimension name}
\usage{
sql_get_dimension_id_from_table_id_and_dimension(
  table_id,
  dimension,
  con,
  schema = "test_platform"
)
}
\arguments{
\item{table_id}{Integer table identifier}

\item{dimension}{Character string dimension name}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Integer dimension ID or NULL if not found
}
\description{
Get dimension ID for table ID and dimension name
}
