% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_initials_from_author_name}
\alias{sql_get_initials_from_author_name}
\title{Get author initials from author name}
\usage{
sql_get_initials_from_author_name(author_name, con, schema = "test_platform")
}
\arguments{
\item{author_name}{Character string author name}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing author initials or NA if not found
}
\description{
Get author initials from author name
}
