% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_interval_from_vintage}
\alias{sql_get_interval_from_vintage}
\title{Get interval ID from vintage ID}
\usage{
sql_get_interval_from_vintage(con, vintage_id, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{vintage_id}{Integer vintage identifier}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing interval ID or NULL if not found
}
\description{
Get interval ID from vintage ID
}
