% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_last_publication_date_from_table_id}
\alias{sql_get_last_publication_date_from_table_id}
\title{Get publication dates for table ID}
\usage{
sql_get_last_publication_date_from_table_id(
  table_id,
  con,
  schema = "test_platform"
)
}
\arguments{
\item{table_id}{Integer table identifier}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Vector of publication timestamps or NULL if none found
}
\description{
Get publication dates for table ID
}
