% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_level_value_from_text}
\alias{sql_get_level_value_from_text}
\title{Get dimension level value from level text}
\usage{
sql_get_level_value_from_text(tab_dim_id, level_text, con, schema = "platform")
}
\arguments{
\item{tab_dim_id}{Integer table dimension identifier}

\item{level_text}{Character string level text}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string level value or NULL if not found
}
\description{
Get dimension level value from level text
}
