% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_levels_from_dimension_id}
\alias{sql_get_levels_from_dimension_id}
\title{Get dimension levels from dimension ID}
\usage{
sql_get_levels_from_dimension_id(tab_dim_id, con, schema = "test_platform")
}
\arguments{
\item{tab_dim_id}{Integer dimension identifier}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame containing dimension levels or NULL if none found
}
\description{
Get dimension levels from dimension ID
}
