% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_max_category_id_for_source}
\alias{sql_get_max_category_id_for_source}
\title{Get maximum category ID for source ID}
\usage{
sql_get_max_category_id_for_source(source_id, con, schema = "platform")
}
\arguments{
\item{source_id}{Integer source identifier}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Integer maximum category ID or 0 if none found
}
\description{
Get maximum category ID for source ID
}
