% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_non_time_dimensions_from_table_id}
\alias{sql_get_non_time_dimensions_from_table_id}
\title{Get non-time dimensions from table ID}
\usage{
sql_get_non_time_dimensions_from_table_id(
  table_id,
  con,
  schema = "test_platform"
)
}
\arguments{
\item{table_id}{Integer table identifier}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame containing dimension names and IDs, or NULL if none found
}
\description{
Get non-time dimensions from table ID
}
