% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_series_id_from_series_code}
\alias{sql_get_series_id_from_series_code}
\title{Get series ID from series code}
\usage{
sql_get_series_id_from_series_code(series_code, con, schema = "platform")
}
\arguments{
\item{series_code}{Character string series code}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Numeric series ID or NA if not found
}
\description{
Get series ID from series code
}
