% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_series_ids_from_table_id}
\alias{sql_get_series_ids_from_table_id}
\title{Get series IDs from table ID}
\usage{
sql_get_series_ids_from_table_id(table_id, con, schema = "platform")
}
\arguments{
\item{table_id}{Integer table identifier}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame of series IDs or NULL if none found
}
\description{
Get series IDs from table ID
}
