% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_series_name_from_series_code}
\alias{sql_get_series_name_from_series_code}
\title{Get series name from series code}
\usage{
sql_get_series_name_from_series_code(
  con,
  series_code,
  schema = "test_platform"
)
}
\arguments{
\item{con}{Database connection object}

\item{series_code}{Character string series code}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing series name or NULL if not found
}
\description{
Get series name from series code
}
