% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_table_id_form_series_id}
\alias{sql_get_table_id_form_series_id}
\title{Get latest datapoints for table with vintages and all}
\usage{
sql_get_table_id_form_series_id(series_id, con, schema = "platform")
}
\arguments{
\item{series_id}{what it says on the tin}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with table IDs and codes
}
\description{
Get latest datapoints for table with vintages and all
}
