% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_table_id_from_table_code}
\alias{sql_get_table_id_from_table_code}
\title{Get table ID from table code}
\usage{
sql_get_table_id_from_table_code(con, table_code, schema = "test_platform")
}
\arguments{
\item{con}{Database connection object}

\item{table_code}{Character string with table code}

\item{schema}{Character string specifying the database schema}
}
\value{
Numeric table ID or NA if not found
}
\description{
Get table ID from table code
}
