% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_table_info}
\alias{sql_get_table_info}
\title{Get table information by table id}
\usage{
sql_get_table_info(table_id, con, schema = "platform")
}
\arguments{
\item{table_id}{Integer specifying the table id}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with table information or NULL if no table found
}
\description{
Get table information by table id
}
