% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_table_name_from_series}
\alias{sql_get_table_name_from_series}
\title{Get table name from series ID}
\usage{
sql_get_table_name_from_series(con, series_id, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{series_id}{Integer series identifier}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing table name or NULL if not found
}
\description{
Get table name from series ID
}
