% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_tables_from_source}
\alias{sql_get_tables_from_source}
\title{Get tables filtered by source_id and update status}
\usage{
sql_get_tables_from_source(
  con,
  schema = "platform",
  source_id = NULL,
  update = NULL
)
}
\arguments{
\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}

\item{source_id}{Optional integer specifying the source ID. If NULL, returns tables from all sources.}

\item{update}{Optional boolean specifying the update status. If NULL, returns tables with any update status.}
}
\value{
Data frame with table IDs and codes
}
\description{
Get tables filtered by source_id and update status
}
