% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_tables_with_keep_vintage}
\alias{sql_get_tables_with_keep_vintage}
\title{Get tables with keep vintage true or false}
\usage{
sql_get_tables_with_keep_vintage(keep, con, schema = "platform")
}
\arguments{
\item{keep}{logical value for keep vintage}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame containing table information with id, code and name or NULL if none found
}
\description{
Get tables with keep vintage true or false
}
