% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_time_dimension_from_table_code}
\alias{sql_get_time_dimension_from_table_code}
\title{Get time dimension name from table code}
\usage{
sql_get_time_dimension_from_table_code(table_code, con, schema = "platform")
}
\arguments{
\item{table_code}{Character string table code}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string time dimension name or NULL if not found
}
\description{
Get time dimension name from table code
}
