% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_unit_from_series}
\alias{sql_get_unit_from_series}
\title{Get unit name from series id}
\usage{
sql_get_unit_from_series(con, series_id, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{series_id}{Integer series identifier}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing unit name or NULL if not found
}
\description{
Get unit name from series id
}
