% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_unit_from_vintage}
\alias{sql_get_unit_from_vintage}
\title{Get unit name from vintage ID}
\usage{
sql_get_unit_from_vintage(con, vintage_id, schema = "test_platform")
}
\arguments{
\item{con}{Database connection object}

\item{vintage_id}{Integer vintage identifier}

\item{schema}{Character string specifying the database schema}
}
\value{
Character string containing unit name or NULL if not found
}
\description{
Get unit name from vintage ID
}
