% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_vintage_from_series}
\alias{sql_get_vintage_from_series}
\title{Get vintage ID from series ID}
\usage{
sql_get_vintage_from_series(
  con,
  series_id,
  date_valid = NULL,
  schema = "test_platform"
)
}
\arguments{
\item{con}{Database connection object}

\item{series_id}{Integer series identifier, single or vector or column}

\item{date_valid}{Timestamp when the vintage was valid (optional)}

\item{schema}{Character string specifying the database schema}
}
\value{
Integer vintage ID or NULL if not found
}
\description{
Get vintage ID from series ID
}
