% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_vintage_from_series_code}
\alias{sql_get_vintage_from_series_code}
\title{Get vintage ID from series code}
\usage{
sql_get_vintage_from_series_code(
  con,
  series_code,
  date_valid = NULL,
  schema = "platform"
)
}
\arguments{
\item{con}{Database connection object}

\item{series_code}{Character string series code}

\item{date_valid}{Timestamp when the vintage was valid (optional)}

\item{schema}{Character string specifying the database schema}
}
\value{
Integer vintage ID or NULL if not found
}
\description{
Get vintage ID from series code
}
