% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_vintages_from_series}
\alias{sql_get_vintages_from_series}
\title{Get all vintages for a series}
\usage{
sql_get_vintages_from_series(series_id, con, schema = "platform")
}
\arguments{
\item{series_id}{Integer specifying the series id}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with all vintages for the series, ordered by published date (newest first)
}
\description{
Get all vintages for a series
}
