% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_date_published_from_vintage}
\alias{get_date_published_from_vintage}
\title{Get the publication date from the vintage id}
\usage{
get_date_published_from_vintage(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
POSIXct value with CET timezone
}
\description{
Just does that.
}
\details{
Deprecated: Use `\link[UMARaccessR]{sql_get_date_published_from_vintage}` instead.
}
