% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_email_from_author_initials}
\alias{get_email_from_author_initials}
\title{Get email from author initials}
\usage{
get_email_from_author_initials(initials, con)
}
\arguments{
\item{initials}{author's iniitals}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
character email
}
\description{
Get email from author initials
}
