% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{get_table_id_from_table_code}
\alias{get_table_id_from_table_code}
\title{Get table id from table code}
\usage{
get_table_id_from_table_code(code, con)
}
\arguments{
\item{code}{character string with source name}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
numeric code
}
\description{
Get table id from table code
}
