% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stuff.R
\name{sql_get_vintages_with_hashes_from_series_id}
\alias{sql_get_vintages_with_hashes_from_series_id}
\title{Get category ID from category name}
\usage{
sql_get_vintages_with_hashes_from_series_id(
  series_id,
  con,
  schema = "platform"
)
}
\arguments{
\item{series_id}{series id}

\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}
}
\value{
Data frame with hashes
}
\description{
Get category ID from category name
}
